// WADNR.Web/src/app/shared/generated/model/tag-lookup-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class TagLookupItem { 
    TagID?: number;
    TagName?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface TagLookupItemForm { 
    TagID?: FormControl<number>;
    TagName?: FormControl<string>;
}

export class TagLookupItemFormControls { 
    public static TagID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static TagName = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}