// WADNR.Web/src/app/shared/generated/model/project-tag-save-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectTagSaveRequest { 
    TagIDs?: Array<number> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectTagSaveRequestForm { 
    TagIDs?: FormControl<Array<number>>;
}

export class ProjectTagSaveRequestFormControls { 
    public static TagIDs = (value: FormControlState<Array<number>> | Array<number> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<number>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}